+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.1.0] - 2025-03-24
+
+- Introduced create fetch bloc;
+- Introduced create bloc;
+- Introduced create screen;
+- Refactored create feature;
+- Added publisher;
 
 ## [0.0.1] - 2025-03-24